[feature/issue_136] moved CheckUtil class to util module
