+++ dimmer-core/src/main/java/com/github/cloudyrock/dimmer/DimmerFeatureConfigurable.java
@@ -42,7 +42,7 @@ abstract class DimmerFeatureConfigurable<RUNNER extends DimmerFeatureConfigurabl
      * @return A new immutable instance of a DimmerFeatureConfigurable with the current configuration applied.
      */
     public RUNNER environments(String... environments) {
-        Util.checkArgumentNullEmpty(environments, "environments");
+        CheckUtil.checkArgumentNullEmpty(environments, "environments");
         final List<String> envs = Arrays.asList(environments);
         return newInstance(envs, configMetadata, defaultExceptionType, dimmerConfigReader);
     }
@@ -263,7 +263,7 @@ abstract class DimmerFeatureConfigurable<RUNNER extends DimmerFeatureConfigurabl
      */
     public RUNNER setDefaultExceptionType(
             Class<? extends RuntimeException> newDefaultExceptionType) {
-        Util.checkArgumentNullEmpty(newDefaultExceptionType, "defaultExceptionType");
+        CheckUtil.checkArgumentNullEmpty(newDefaultExceptionType, "defaultExceptionType");
         ExceptionUtil.checkExceptionConstructorType(newDefaultExceptionType);
         return newInstance(environments, configMetadata, newDefaultExceptionType, dimmerConfigReader);
     }
