+++ dimmer-util/src/main/java/com/github/cloudyrock/dimmer/CheckUtil.java
@@ -1,8 +1,8 @@
-package com.github.cloudyrock.dimmer.config.util;
+package com.github.cloudyrock.dimmer;
 
 import java.util.Collection;
 
-public final class Util {
+public final class CheckUtil {
 
     public static void checkNullEmpty(Object obj, String name) {
         if (obj == null || "".equals(obj) || (obj instanceof Collection && ((Collection) obj).isEmpty())) {
